Removed All Old Code
Removed all the old code from the forked copy of the repository. Made me
realize just how little is used from the old project, and how much more
functionality I have in my program comparitivly. I understand why some
of it was written how it was, but it seems overly complex to achive a
very simple result. Maybe I'm the one going about it the wrong way
though. It's possible that the old code was far less error prone, and
more easily controlled when new features were to be added.

Oh well, that's why this is a whole new branch
