--- D360/Bindings/D3Command.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace D360.Bindings
-{
-    public enum D3Command
-    {
-        PrimarySkill,
-        SecondarySkill,
-        ActionBarSkill1,
-        ActionBarSkill2,
-        ActionBarSkill3,
-        ActionBarSkill4,
-        Inventory,
-        Map,
-        Potion,
-        TownPortal,
-        GameMenu,
-        ForceMove,
-        HoldPosition,
-        LeftClick,
-        RightClick,
-    }
-}
