--- D360/Bindings/MouseMove.cs
@@ -1,11 +1,0 @@
-﻿using D360.Types;
-
-namespace D360.Bindings
-{
-    public class MouseMove
-    {
-        public CommandTarget commandTarget;
-        public MouseMoveType moveType;
-        public UIntVector moveScale;
-    }
-}
