--- D360/Bindings/StateChange.cs
@@ -1,10 +1,0 @@
-﻿using D360.Types;
-
-namespace D360.Bindings
-{
-    public class StateChange
-    {
-        public InputMode newMode;
-        public bool toggle;
-    }
-}
