--- D360/Bindings/TriggerState.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace D360.Bindings
-{
-    public enum ControllerButtonState { Down, Up }
-}
