--- D360/Commands/Command.cs
@@ -1,19 +1,0 @@
-﻿using D360.Utility;
-using D360.Types;
-
-namespace D360.Commands
-{
-    public class Command
-    {
-        public InputMode applicableMode = InputMode.None;
-        public CommandTarget target = CommandTarget.Cursor;
-        
-        public virtual bool Execute(ref OldControllerState state)
-        {
-            if (applicableMode == InputMode.None)
-                return false;
-
-            return (applicableMode == state.inputMode) || (applicableMode == InputMode.All);
-        }
-    }
-}
