+++ D360/Types/SpecialAction.cs
@@ -1,0 +1,10 @@
+﻿namespace D360.Types
+{
+    public enum SpecialAction
+    {
+        SwitchStickMode,
+        InventoryMode,
+        Loot,
+        PauseApplication
+    }
+}
